Homologado

+++ src/main/java/com/anatonelly/freteexpress/controller/FreteMotoristaController.java
@@ -37,13 +37,12 @@ public class FreteMotoristaController {
     }
 
     @PostMapping("/motorista/solicitar-frete")
-    public String solicitarFrete(@RequestParam Integer freteId, Model model) { // <<<<< CORRIGIDO AQUI: Long para Integer
+    public String solicitarFrete(@RequestParam Long freteId, Model model) {
         try {
             // Simulação: Motorista logado com ID 1 (ajuste conforme autenticação)
-            // Use 1, sem 'L' para que seja Integer
-            Optional<Motorista> motoristaOpt = motoristaRepository.findById(1); // <<<<< CORRIGIDO AQUI: 1L para 1
+            Optional<Motorista> motoristaOpt = motoristaRepository.findById(1L);
             if (motoristaOpt.isPresent()) {
-                if (freteService.solicitarFrete(freteId, 1)) { // <<<<< CORRIGIDO AQUI: 1L para 1
+                if (freteService.solicitarFrete(freteId, 1L)) {
                     model.addAttribute("message", "Frete solicitado com sucesso!");
                 } else {
                     model.addAttribute("error", "Frete não disponível para solicitação.");
